requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
dateparser==1.1.8
demjson3==3.0.6
facebook-scraper==0.2.59
fake-useragent==1.1.3
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
numpy==1.25.0
pandas==2.0.3
parse==1.19.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
w3lib==2.1.1
waitress==2.1.2
websockets==10.4
Werkzeug==2.3.6
zipp==3.15.0
